+++ src/syntax.rs
@@ -56,8 +56,13 @@ pub fn highlight(line: &str) -> String {
                 is_colored = true;
                 result.push_str(k);
             }
-            (TokenKind::StringRaw, s) => {
+            (TokenKind::Regex, s) => {
                 result.push_str("\x1b[38;5;203m");
+                is_colored = true;
+                result.push_str(s);
+            }
+            (TokenKind::StringRaw, s) => {
+                result.push_str("\x1b[38;5;204m");
                 is_colored = true;
                 result.push_str(s);
             }